# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/error": {
            "get": {
                "description": "自定义错误返回",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "公共相关"
                ],
                "summary": "自定义错误返回",
                "parameters": [
                    {
                        "type": "string",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "msg",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": \"0000\", msg: \"string\", data: \"string\"}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\": \"\", msg: \"\", data: \"\"}",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}